Added step to alter, sort, reindex the filtered bams.
ngsplot uses reference genomes with the 'chr' notation at the beginning of the chromosomes. Using Ensembl genome builds, as I almost always do, means that the bam files must be altered, sorted, and reindexed to work with ngsplot. An option in the config file, if set to Y, will change in input of the ngsplot step to account for this.