Added parsing step to pepr output files.
PePr output files contain ###.0, in stead of integers. Many annotation tools complain about that, so I fixed that and added 'chr' as a prefix to the peaks. There is not a 'chr' containing file and one without the prefix. Either can be used depending on the what the tool might need/like and the genome of interest.